network: move parser to separate file
