+++ network/ports.go
@@ -3,8 +3,6 @@ package network
 import (
 	"errors"
 	"net"
-	"strconv"
-	"strings"
 
 	"go.uber.org/zap"
 )
@@ -38,33 +36,4 @@ func (reg *Registry) AssignPort() (string, error) {
 		return p, nil
 	}
 	return "", errors.New("no available port found")
-}
-
-func parsePorts(portRanges []string) []string {
-	allPorts := make([]string, 0)
-	for _, r := range portRanges {
-		if strings.Contains(r, "-") {
-			// is range
-			ports := strings.Split(r, "-")
-			lower, err := strconv.Atoi(ports[0])
-			if err != nil {
-				continue
-			}
-			upper, err := strconv.Atoi(ports[1])
-			if err != nil {
-				continue
-			}
-			for p := lower; p <= upper; p++ {
-				allPorts = append(allPorts, strconv.Itoa(p))
-			}
-		} else {
-			// check if int
-			if _, err := strconv.Atoi(r); err != nil {
-				continue
-			}
-			// is single port
-			allPorts = append(allPorts, r)
-		}
-	}
-	return allPorts
 }
