+++ network/ports_test.go
@@ -2,7 +2,6 @@ package network
 
 import (
 	"net"
-	"reflect"
 	"testing"
 
 	"github.com/RTradeLtd/ipfs-orchestrator/log"
@@ -27,6 +26,7 @@ func TestRegistry_AssignPort(t *testing.T) {
 		want    string
 		wantErr bool
 	}{
+		{"nil ports", fields{nil}, "", true},
 		{"no ports", fields{[]string{}}, "", true},
 		{"no available port", fields{[]string{"9999"}}, "", true},
 		{"available port", fields{[]string{"9998"}}, "9998", false},
@@ -42,35 +42,6 @@ func TestRegistry_AssignPort(t *testing.T) {
 			}
 			if got != tt.want {
 				t.Errorf("Registry.AssignPort() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_parsePorts(t *testing.T) {
-	type args struct {
-		portRanges []string
-	}
-	tests := []struct {
-		name string
-		args args
-		want []string
-	}{
-		{"bad input", args{nil}, []string{}},
-		{"non-int port", args{[]string{"abcde"}}, []string{}},
-		{"non-int port range", args{[]string{"abcde-fghijk"}}, []string{}},
-		{"mixed non-int port range", args{[]string{"7000-abcde"}}, []string{}},
-		{"bad range", args{[]string{"8000-7999"}}, []string{}},
-		{"single port", args{[]string{"8000"}}, []string{"8000"}},
-		{"multi port", args{[]string{"8000", "8001"}}, []string{"8000", "8001"}},
-		{"port range", args{[]string{"8000-8003"}}, []string{"8000", "8001", "8002", "8003"}},
-		{"multi port range", args{[]string{"8000-8001", "8002-8003"}}, []string{"8000", "8001", "8002", "8003"}},
-		{"mix", args{[]string{"8000", "8002-8003"}}, []string{"8000", "8002", "8003"}},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := parsePorts(tt.args.portRanges); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("parsePorts() = %v, want %v", got, tt.want)
 			}
 		})
 	}
